Added basic functionality for correlation surface, not fully implemented, but has no impact as long as CORRELATION_DIRECTION keyword is not in use. Check in now for coordination with new SegY reader.
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@79 176de4d3-4747-0410-aa96-47f6d6634214

+++ lib/lib_matr.c
@@ -278,7 +278,7 @@ void lib_matrAXeqBMatR(int i_dim, double **i_mat, double **x_mat, int n)
 
   int l_i, l_j;
   double l_x;
-  float help;
+  double help;
   int i;
   for(i=0;i<n;i++)
   {
